+++ packages/EasyActivity/src/Common/Object/FullySerializableObjectInterface.php
@@ -1,11 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyActivity\Common\Serializer;
+namespace EonX\EasyActivity\Common\Object;
 
 /**
  * All properties of the object implementing this interface will be allowed to be serialized.
  */
-interface AllowAllPropertiesSerializerInterface
+interface FullySerializableObjectInterface
 {
 }
